server: reject device-supplied column names containing weird characters
The JSON packet format allows a device to specify the column names
that will be used in the database. These column names are used to
construct the parameterized query string that is supplied to the
database cursor's execute() method.

Inside create_query(), use re.search() to raise an exception upon
receiving a data packet containing a column name with shady
non-alphanumeric characters, e.g.:

    ") VALUES (12345); DROP TABLE outdoor_env;--"