Added DTR, RSSI, and Xbee sleep pin to Dragonfruit (#58)
Set the DTR pin as output and as logical low. Since the XBee is in router mode, the output doesn't matter so we can drive this signal to LOW.

Set RSSI pin as input. Since the RSSI pin is set to output on XBee, we correspondingly set the RSSI pin to input on the MCU.

Set Xbee_On/Sleep pin as input to idle the XBee. Similarly, since the Xbee_On/Sleep pin is set to output on the XBee, we set the pin to input on the MCU.