Fixes to dragonfruit solar panel and battery pins (#61)
On Dragonfruit, there were already connections for solar panel and battery readings made on version 2. Previously it was assumed there was no connections to the MCU so a wire was inserted thus VBATT was connected to A0 and SOLAR_TEST was connected to A6. Since there are already connections, the wire on the board was removed and the pins are now correspondingly changed to A7 for VBATT and A0 for SOLAR_TEST.